Migrate to MVC: Rewrite `ALL /auth`
